Streams audio more efficiently. Fixed media photo bug in notification. Showing media_screen in tabs.

+++ app/src/main/java/com/roland/android/odiyo/model/Music.kt
@@ -11,9 +11,9 @@ data class Music(
 	val name: String,
 	val title: String,
 	val artist: String,
-	private val time: Int,
+	private val time: Long,
 	val thumbnail: Bitmap?
 ) {
 	@RequiresApi(Build.VERSION_CODES.Q)
-	val duration = time.toLong().time
+	val duration = time.time
 }
